+++ frontend/src/pages/Login/index.ts
@@ -1,0 +1,2 @@
+import Login from './Login' 
+export default Login
